# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  description: string;
  accent?: string;
  toolCount: number;
}

export interface CategoryInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  description: string;
  accent?: string;
}

export interface CategoryUpdate {
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  description?: string;
  accent?: string;
}

export interface Tool {
  id: number;
  name: string;
  slug: string;
  shortDescription: string;
  description: string;
  howItWorks: string;
  bestFor?: string;
  websiteUrl: string;
  /** @nullable */
  logoUrl?: string | null;
  /** @nullable */
  previewUrl?: string | null;
  categoryId: number;
  categoryName: string;
  categorySlug?: string;
  pricing: string;
  featured: boolean;
  status: string;
  createdAt: string;
}

export interface ToolInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  shortDescription: string;
  /** @minLength 1 */
  description: string;
  /** @minLength 1 */
  howItWorks: string;
  bestFor?: string;
  websiteUrl: string;
  /** @nullable */
  logoUrl?: string | null;
  /** @nullable */
  previewUrl?: string | null;
  categoryId: number;
  /** @minLength 1 */
  pricing: string;
  featured?: boolean;
  status?: string;
}

export interface ToolUpdate {
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  shortDescription?: string;
  /** @minLength 1 */
  description?: string;
  /** @minLength 1 */
  howItWorks?: string;
  bestFor?: string;
  websiteUrl?: string;
  /** @nullable */
  logoUrl?: string | null;
  /** @nullable */
  previewUrl?: string | null;
  categoryId?: number;
  /** @minLength 1 */
  pricing?: string;
  featured?: boolean;
  status?: string;
}

export type ListToolsParams = {
search?: string;
category?: string;
featured?: boolean;
status?: string;
};