# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: tools
    description: Tool discovery and management
  - name: categories
    description: Tool category management
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /tools:
    get:
      operationId: listTools
      tags: [tools]
      summary: List tools
      parameters:
        - name: search
          in: query
          required: false
          schema: { type: string }
        - name: category
          in: query
          required: false
          schema: { type: string }
        - name: featured
          in: query
          required: false
          schema: { type: boolean }
        - name: status
          in: query
          required: false
          schema: { type: string }
      responses:
        "200":
          description: Tools
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tool"
    post:
      operationId: createTool
      tags: [tools]
      summary: Create a tool
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ToolInput"
      responses:
        "201":
          description: Created tool
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tool"
  /tools/{id}:
    get:
      operationId: getTool
      tags: [tools]
      summary: Get a tool by id
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Tool
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tool"
        "404":
          description: Tool not found
    patch:
      operationId: updateTool
      tags: [tools]
      summary: Update a tool
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ToolUpdate"
      responses:
        "200":
          description: Updated tool
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tool"
        "404":
          description: Tool not found
    delete:
      operationId: deleteTool
      tags: [tools]
      summary: Delete a tool
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Tool deleted
        "404":
          description: Tool not found
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List categories
      responses:
        "200":
          description: Categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Created category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    patch:
      operationId: updateCategory
      tags: [categories]
      summary: Update a category
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryUpdate"
      responses:
        "200":
          description: Updated category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "404":
          description: Category not found
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete a category
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Category deleted
        "404":
          description: Category not found
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Category:
      type: object
      required: [id, name, slug, description, toolCount]
      properties:
        id: { type: integer }
        name: { type: string }
        slug: { type: string }
        description: { type: string }
        accent: { type: string }
        toolCount: { type: integer }
    CategoryInput:
      type: object
      required: [name, description]
      properties:
        name: { type: string, minLength: 1 }
        description: { type: string, minLength: 1 }
        accent: { type: string }
    CategoryUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        description: { type: string, minLength: 1 }
        accent: { type: string }
    Tool:
      type: object
      required: [id, name, slug, shortDescription, description, howItWorks, websiteUrl, categoryId, categoryName, pricing, featured, status, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        slug: { type: string }
        shortDescription: { type: string }
        description: { type: string }
        howItWorks: { type: string }
        bestFor: { type: string }
        websiteUrl: { type: string, format: uri }
        logoUrl: { type: ["string", "null"], format: uri }
        previewUrl: { type: ["string", "null"] }
        categoryId: { type: integer }
        categoryName: { type: string }
        categorySlug: { type: string }
        pricing: { type: string }
        featured: { type: boolean }
        status: { type: string }
        createdAt: { type: string, format: date-time }
    ToolInput:
      type: object
      required: [name, shortDescription, description, howItWorks, websiteUrl, categoryId, pricing]
      properties:
        name: { type: string, minLength: 1 }
        shortDescription: { type: string, minLength: 1 }
        description: { type: string, minLength: 1 }
        howItWorks: { type: string, minLength: 1 }
        bestFor: { type: string }
        websiteUrl: { type: string, format: uri }
        logoUrl: { type: ["string", "null"], format: uri }
        previewUrl: { type: ["string", "null"] }
        categoryId: { type: integer }
        pricing: { type: string, minLength: 1 }
        featured: { type: boolean }
        status: { type: string }
    ToolUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        shortDescription: { type: string, minLength: 1 }
        description: { type: string, minLength: 1 }
        howItWorks: { type: string, minLength: 1 }
        bestFor: { type: string }
        websiteUrl: { type: string, format: uri }
        logoUrl: { type: ["string", "null"], format: uri }
        previewUrl: { type: ["string", "null"] }
        categoryId: { type: integer }
        pricing: { type: string, minLength: 1 }
        featured: { type: boolean }
        status: { type: string }