# lib/api-zod/src/generated/types/categoryUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CategoryUpdate {
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  description?: string;
  accent?: string;
}