# lib/api-zod/src/generated/types/tool.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Tool {
  id: number;
  name: string;
  slug: string;
  shortDescription: string;
  description: string;
  howItWorks: string;
  bestFor?: string;
  websiteUrl: string;
  /** @nullable */
  logoUrl?: string | null;
  /** @nullable */
  previewUrl?: string | null;
  categoryId: number;
  categoryName: string;
  categorySlug?: string;
  pricing: string;
  featured: boolean;
  status: string;
  createdAt: Date;
}