# lib/api-zod/src/generated/types/toolUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ToolUpdate {
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  shortDescription?: string;
  /** @minLength 1 */
  description?: string;
  /** @minLength 1 */
  howItWorks?: string;
  bestFor?: string;
  websiteUrl?: string;
  /** @nullable */
  logoUrl?: string | null;
  /** @nullable */
  previewUrl?: string | null;
  categoryId?: number;
  /** @minLength 1 */
  pricing?: string;
  featured?: boolean;
  status?: string;
}